Add data validation

+++ db.py
@@ -1,4 +1,8 @@
-import sqlite3
+import sqlite3, re
+
+
+class UserDoesNotExistError(Exception):
+    pass
 
 
 class BotDB:
@@ -15,6 +19,8 @@ class BotDB:
 
     def get_user_id(self, user_id):
         """Получение пользовательского ID в БД"""
+        if not self.user_exists(user_id):
+            raise UserDoesNotExistError("User does not exist")
         result = self.cursor.execute("SELECT Id FROM users WHERE userId = ?", (user_id,))
         return result.fetchone()[0]
 
@@ -26,23 +32,34 @@ class BotDB:
 
     def update_location(self, user_id, city):
         """Смена города для оповещения о погоде"""
+        if not self.user_exists(user_id):
+            raise UserDoesNotExistError("User does not exist")
         self.cursor.execute("UPDATE locations SET city = ? WHERE userId = ?", (city, self.get_user_id(user_id),))
         return self.connect.commit()
 
     def update_notification_time(self, user_id, time):
-        """Смена времени оповещения о погоде"""
-        self.cursor.execute("UPDATE locations SET notificationTime = ? WHERE userId = ?",
-                            (time, self.get_user_id(user_id),))
-        return self.connect.commit()
+        """Смена времени оповещения о погоде в формате HH:MM"""
+        if not self.user_exists(user_id):
+            raise UserDoesNotExistError("User does not exist")
+        if bool(re.match(r"[0-2][\d]:[0-5][\d]", time)):
+            self.cursor.execute("UPDATE locations SET notificationTime = ? WHERE userId = ?",
+                                (time, self.get_user_id(user_id),))
+            return self.connect.commit()
+        else:
+            raise ValueError("Time format error")
 
     def update_rating(self, user_id, rating):
         """Обновить рейтинг пользователя"""
+        if not self.user_exists(user_id):
+            raise UserDoesNotExistError("User does not exist")
         self.cursor.execute("UPDATE users SET rating = ? WHERE Id = ?",
                             (rating, self.get_user_id(user_id)))
         return self.connect.commit()
 
     def get_location(self, user_id):
         """Получение города и времени для погодного оповещения"""
+        if not self.user_exists(user_id):
+            raise UserDoesNotExistError("User does not exist")
         self.cursor.execute("SELECT city, notificationTime FROM locations WHERE userId = ?",
                             (self.get_user_id(user_id),))
         return self.cursor.fetchall()
@@ -54,6 +71,8 @@ class BotDB:
 
     def delete_user(self, user_id):
         """Удаление пользователя из БД"""
+        if not self.user_exists(user_id):
+            raise UserDoesNotExistError("User does not exist")
         self.cursor.execute("DELETE FROM locations WHERE userId = ?", (self.get_user_id(user_id),))
         self.cursor.execute("DELETE FROM users WHERE userId = ?", (user_id,))
         return self.connect.commit()
